add environment variable DJ_SUPPORT_FILEPATH_MANAGEMENT to enable/disable the filepath datatype
