+++ datajoint/version.py
@@ -1,3 +1,3 @@
-__version__ = "0.12.dev7"
+__version__ = "0.12.dev8"
 
 assert len(__version__) <= 10  # The log table limits version to the 10 characters
