+++ tests/test_adapted_attributes.py
@@ -5,12 +5,9 @@ from nose.tools import assert_true, assert_equal
 from . import schema_adapted as adapted
 from .schema_adapted import graph
 
-from datajoint import attribute_adapter
-
-
 
 def test_adapted_type():
-    attribute_adapter._switch_adapated_types(True)
+    dj.errors._switch_adapted_types(True)
     c = adapted.Connectivity()
     graphs = [nx.lollipop_graph(4, 2), nx.star_graph(5), nx.barbell_graph(3, 1), nx.cycle_graph(5)]
     c.insert((i, g) for i, g in enumerate(graphs))
@@ -20,7 +17,7 @@ def test_adapted_type():
         assert_equal(len(g1.edges), len(g2.edges))
         assert_true(0 == len(nx.symmetric_difference(g1, g2).edges))
     c.delete()
-    attribute_adapter._switch_adapated_types(False)
+    dj.errors._switch_adapted_types(False)
 
 
 # test spawned classes
@@ -29,7 +26,7 @@ local_schema.spawn_missing_classes()
 
 
 def test_adapted_spawned():
-    attribute_adapter._switch_adapated_types(True)
+    dj.errors._switch_adapted_types(True)
     c = Connectivity()  # a spawned class
     graphs = [nx.lollipop_graph(4, 2), nx.star_graph(5), nx.barbell_graph(3, 1), nx.cycle_graph(5)]
     c.insert((i, g) for i, g in enumerate(graphs))
@@ -39,7 +36,7 @@ def test_adapted_spawned():
         assert_equal(len(g1.edges), len(g2.edges))
         assert_true(0 == len(nx.symmetric_difference(g1, g2).edges))
     c.delete()
-    attribute_adapter._switch_adapated_types(False)
+    dj.errors._switch_adapted_types(False)
 
 
 # test with virtual module
@@ -47,7 +44,7 @@ virtual_module = dj.create_virtual_module('virtual_module', adapted.schema_name,
 
 
 def test_adapted_virtual():
-    attribute_adapter._switch_adapated_types(True)
+    dj.errors._switch_adapted_types(True)
     c = virtual_module.Connectivity()
     graphs = [nx.lollipop_graph(4, 2), nx.star_graph(5), nx.barbell_graph(3, 1), nx.cycle_graph(5)]
     c.insert((i, g) for i, g in enumerate(graphs))
@@ -61,7 +58,7 @@ def test_adapted_virtual():
             assert_equal(len(g1.edges), len(g2.edges))
             assert_true(0 == len(nx.symmetric_difference(g1, g2).edges))
     c.delete()
-    attribute_adapter._switch_adapated_types(False)
+    dj.errors._switch_adapted_types(False)
